"Sign up with profile picture and multiple categories as well as event radius choice applied"

+++ app/src/main/java/com/example/buckit/fragments/EventsExploreFragment.java
@@ -65,8 +65,8 @@ public class EventsExploreFragment extends Fragment implements CardStack.CardEve
     public final static String KEY_DEFAULT_RADIUS = "100mi";
     public View popupView;
     public SwipeCardAdapter swipe_card_adapter;
-    public Double latitude;
-    public Double longitude;
+    public float latitude;
+    public float longitude;
     public HashMap<Integer, Event> eventsList;
     @BindView(R.id.rvEvents) public CardStack rvEvents;
     @BindView(R.id.ivBlur) public ImageView blur;
@@ -108,8 +108,8 @@ public class EventsExploreFragment extends Fragment implements CardStack.CardEve
         rvEvents.setAdapter(swipe_card_adapter);
         rvEvents.setListener(this);
         if (getArguments() != null){
-            latitude = Double.valueOf(getArguments().getFloat(LAT_KEY));
-            longitude = Double.valueOf(getArguments().getFloat(LONG_KEY));
+            latitude = getArguments().getFloat(LAT_KEY);
+            longitude = getArguments().getFloat(LONG_KEY);
             getEvents(user);
         }
         mCheck.bringToFront();
